Changing package to com.xwiki.googleapps. Paul
