+++ api/src/main/java/com/xwiki/googleapps/GoogleAppsScriptService.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.apps.googleapps;
+package com.xwiki.googleapps;
 
 import java.util.Date;
 import java.util.List;
