+++ api/src/main/resources/META-INF/components.txt
@@ -1,4 +1,4 @@
-org.xwiki.apps.googleapps.internal.GoogleAppsManagerImpl
-org.xwiki.apps.googleapps.internal.CookieAuthenticationPersistenceImpl
-org.xwiki.apps.googleapps.GoogleAppsScriptService
-org.xwiki.apps.googleapps.internal.GoogleAppsAuthServiceImpl
+com.xwiki.googleapps.internal.GoogleAppsManagerImpl
+com.xwiki.googleapps.internal.CookieAuthenticationPersistenceImpl
+com.xwiki.googleapps.internal.GoogleAppsAuthServiceImpl
+com.xwiki.googleapps.GoogleAppsScriptService
